Code changes for:
1) Adding Services to the category.
2) Updated project.sql
3) Implemented a new feature to the application
	--> On very first deployment of application with no data in DB, we need
to create ADMIN user first.
	--> So with help of registration page we will create ADMIN user, but we
need to restrict the USERS created from registration page as application
has the feature to add STAFF where automatically users we be created
when staff is added.
	--> With current changes we are restricting the max count of users that
are created from registration page is 3.

4) Some indentation changes.

+++ src/main/java/com/springbootexample/dao/StaffRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.springbootexample.model.Staff;
 
-public interface StaffRepository extends JpaRepository<Staff, Integer>{
+public interface StaffRepository extends JpaRepository<Staff, Long>{
 }
