+++ src/main/java/com/springbootexample/model/Role.java
@@ -13,14 +13,14 @@ public class Role extends TimeStampClass {
 	@Id
     @GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name="role_id")
-	private int id;
+	private Long id;
 	@Column(name="role")
 	private String role;
 	
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getRole() {
